Module Current_rpc

module Job : sig ... end

Client-side API to contact a job service.

module Engine : sig ... end

Client-side API to contact an engine service.

module Impl (Current : sig ... end) : sig ... end

This is used on the server-side to provide access to the OCurrent Engine. Create an instance of the functor with module Rpc = Current_rpc.Impl(Current). We use a functor here just to avoid having Current_rpc depend on Current, which would be annoying for RPC clients.