Module Current_rpc.Impl

This is used on the server-side to provide access to the OCurrent Engine. Create an instance of the functor with module Rpc = Current_rpc.Impl(Current). We use a functor here just to avoid having Current_rpc depend on Current, which would be annoying for RPC clients.

Parameters

module Current : sig ... end

Signature

val job : engine:Current.Engine.t -> Job.id -> Job.t

job ~engine id is a Cap'n Proto job service backed by engine.

val engine : Current.Engine.t -> Engine.t

engine e is a Cap'n Proto engine service backed by e.