Module Current_rpc.Engine

type t = [ `Engine_f0961466d2f9bbf5 ] Capnp_rpc_lwt.Capability.t

val active_jobs : t -> (Job.id list, [> `Capnp of Capnp_rpc.Error.t ]) Stdlib.result Lwt.t

active_jobs t lists the OCurrent jobs that are still being used in the pipeline. This includes completed jobs, as long as OCurrent is still ensuring they are up-to-date.

val job : t -> Job.id -> Job.t

job t id is the job with the given ID. This does not have to be an active job (but only active jobs can be rebuilt). If the job ID is unknown, this operation will resolve to a suitable error.