Module Current_rpc.Job

Client-side API to contact a job service.

type t = [ `Job_8397ef9078537247 ] Capnp_rpc_lwt.Capability.t

type id = string

type status = {

id : id;

description : string;

can_cancel : bool;

can_rebuild : bool;

}

val log : start:int64 -> t -> (string * int64, [> `Capnp of Capnp_rpc.Error.t ]) Stdlib.result Lwt.t

log ~start t returns bytes from the log starting at offset start.

val cancel : t -> (unit, [> `Capnp of Capnp_rpc.Error.t ]) Stdlib.result Lwt.t

val status : t -> (status, [> `Capnp of Capnp_rpc.Error.t ]) Stdlib.result Lwt.t

val rebuild : t -> t

rebuild t requests a rebuild of t and returns the new job.

val approve_early_start : t -> (unit, [> `Capnp of Capnp_rpc.Error.t ]) Stdlib.result Lwt.t